Handle file at root

+++ app.ts
@@ -11,8 +11,7 @@ import packageJson from "./package.json";
 import keypress from "keypress";
 
 const TIMER_END = 10;
-
-keypress(process.stdin);
+const MEDIA_REGEX = /^video\/.*/;
 
 (async () => {
   console.log(
@@ -32,14 +31,14 @@ keypress(process.stdin);
     validate: <T>(
       input: string
     ): boolean | string | Promise<boolean | string> => {
-      return /^(.+)\/([^\/]+)$/.test(input)
+      return /^(.){1, 2}\/|([^\/]+)|\/$/.test(input)
         ? true
         : "The path you input is invalid.";
     },
     default: "../DOWNLOADING",
   });
 
-  const TARGET_DIR = path.resolve(__dirname, target.target);
+  const TARGET_DIR = path.resolve(process.cwd(), target.target);
 
   console.log(
     `The target you specific was \u001b[33m${TARGET_DIR}\u001b[0m\nPlease hit \u001b[32mEnter\u001b[0m to confirm or \u001b[31mCMD + C\u001b[0m to escape`
@@ -50,6 +49,7 @@ keypress(process.stdin);
     let id: NodeJS.Timeout;
     bar.start(TIMER_END, 0);
 
+    keypress(process.stdin);
     process.stdin.on("keypress", (str, key) => {
       if (key.ctrl && key.name === "c") {
         process.exit();
@@ -75,26 +75,33 @@ keypress(process.stdin);
     }, 1000);
   });
 
-  fs.readdirSync(path.resolve(__dirname, "../DOWNLOADING"))
+  fs.readdirSync(TARGET_DIR)
     .filter((dir) => !dir.match(/^\..*/))
     .map((dirname) => {
-      const DIR_PATH = path.join(TARGET_DIR, dirname);
-      fs.readdirSync(DIR_PATH).map((filename) => {
-        const FILE_PATH = path.join(DIR_PATH, filename);
-        if (filename.match(/.*.(js)|(torrent)$/)) {
-          console.log(`remove file ${filename}`);
-          fs.unlinkSync(FILE_PATH);
-        }
+      let DIR_PATH = path.join(TARGET_DIR, dirname);
 
-        const mimeType = mime.getType(FILE_PATH) || "";
-        if (mimeType.match(/^video\/.*/)) {
-          const extension = mime.getExtension(mimeType);
-          fs.renameSync(FILE_PATH, `${DIR_PATH}/${dirname}.${extension}`);
-          console.log(
-            `rename file name: ${filename} -> ${dirname}.${extension}`
-          );
-        }
-      });
+      if (!mime.getType(dirname)) {
+        fs.readdirSync(DIR_PATH).map((filename) => {
+          const FILE_PATH = path.join(DIR_PATH, filename);
+          if (filename.match(/.*.(js)|(torrent)$/)) {
+            console.log(`remove file ${filename}`);
+            fs.unlinkSync(FILE_PATH);
+          }
+
+          const mimeType = mime.getType(FILE_PATH) || "";
+          if (mimeType.match(MEDIA_REGEX)) {
+            const extension = mime.getExtension(mimeType);
+            fs.renameSync(FILE_PATH, `${DIR_PATH}/${dirname}.${extension}`);
+            console.log(
+              `rename file name: ${filename} -> ${dirname}.${extension}`
+            );
+          }
+        });
+      } else if (mime.getType(dirname)?.match(MEDIA_REGEX)) {
+        const REAL_DIR_PATH = DIR_PATH.substring(0, DIR_PATH.lastIndexOf("."));
+        fs.mkdirSync(REAL_DIR_PATH);
+        fs.renameSync(DIR_PATH, path.join(REAL_DIR_PATH, `/${dirname}`));
+      }
     });
   process.exit();
 })();
